+++ backend/ddang/src/main/java/com/ddang/ddang/review/presentation/ReviewController.java
@@ -5,9 +5,9 @@ import com.ddang.ddang.authentication.domain.dto.AuthenticationUserInfo;
 import com.ddang.ddang.image.presentation.util.ImageRelativeUrlFinder;
 import com.ddang.ddang.image.presentation.util.ImageTargetType;
 import com.ddang.ddang.review.application.ReviewService;
-import com.ddang.ddang.review.application.dto.CreateReviewDto;
-import com.ddang.ddang.review.application.dto.ReadReviewDetailDto;
-import com.ddang.ddang.review.application.dto.ReadReviewDto;
+import com.ddang.ddang.review.application.dto.request.CreateReviewDto;
+import com.ddang.ddang.review.application.dto.response.ReadReviewDetailDto;
+import com.ddang.ddang.review.application.dto.response.ReadReviewDto;
 import com.ddang.ddang.review.presentation.dto.request.CreateReviewRequest;
 import com.ddang.ddang.review.presentation.dto.response.ReadReviewDetailResponse;
 import com.ddang.ddang.review.presentation.dto.response.ReadReviewResponse;
