+++ backend/ddang/src/main/java/com/ddang/ddang/review/presentation/dto/response/ReadReviewDetailResponse.java
@@ -1,13 +1,13 @@
 package com.ddang.ddang.review.presentation.dto.response;
 
-import com.ddang.ddang.review.application.dto.ReadReviewDetailDto;
+import com.ddang.ddang.review.application.dto.response.ReadReviewDetailDto;
 import jakarta.annotation.Nullable;
-
 
 public record ReadReviewDetailResponse(@Nullable Float score, @Nullable String content) {
 
     public static ReadReviewDetailResponse from(final ReadReviewDetailDto readReviewDetailDto) {
         final Double nullableScore = readReviewDetailDto.score();
+
         if (nullableScore == null) {
             return new ReadReviewDetailResponse(null, readReviewDetailDto.content());
         }
